(maint) Only create tests/configs directory when needed
Previously the Rakefile would try and create the `tests/configs` directory
when invoked regardless whether it was required or not.  This also caused issues
with other rake tasks which expect to output information to STDOUT but get's
polluted by the error message `mkdir: cannot create directory `tests/configs':
File exists` because that directory is already in the repository.  This commit
moves the directory creation into the reference_tests rake tasks as this is
where it is required.